Serial Tweaks
Fixed an error that was causing the code to not run if it wasn't plugged into a computer (USB serial communication). The code snippet that was causing this was still needed so I just moved it to the setupSerial() method and added a delay. Also made it so the code would not run if and instance of board communication wasn't found.